split click UI from implementation for config cmds
