+++ src/instructlab/cli/config/show.py
